TopNav: Move news into profile menu  (#99535)
* remove news icon from topnav


* TopNav: Move rss feed and kiosk action into profile menu

* Update language keys

* Update

* review fixes

* Update

* Update

---------
