+++ public/app/core/components/AppChrome/TopBar/TopNavBarMenu.tsx
@@ -8,9 +8,10 @@ import { enrichWithInteractionTracking } from '../MegaMenu/utils';
 
 export interface TopNavBarMenuProps {
   node: NavModelItem;
+  children?: React.ReactNode;
 }
 
-export function TopNavBarMenu({ node: nodePlain }: TopNavBarMenuProps) {
+export function TopNavBarMenu({ node: nodePlain, children }: TopNavBarMenuProps) {
   const styles = useStyles2(getStyles);
   const node = enrichWithInteractionTracking(cloneDeep(nodePlain), false);
 
@@ -37,6 +38,7 @@ export function TopNavBarMenu({ node: nodePlain }: TopNavBarMenuProps) {
           <MenuItem icon={item.icon} onClick={item.onClick} label={item.text} key={item.id} />
         );
       })}
+      {children}
     </Menu>
   );
 }
